genai-apps/requirements.txt
pypdf==3.12.2
requests==2.32.3
urllib3==1.26.11
sentence-transformers==2.2.2
transformers[torch]
huggingface_hub[inference]
python-dotenv==1.0.0
streamlit==1.39.0
llama-index-llms-ollama==0.3.4
streamlit==1.39.0
ollama==0.3.3
langchain==0.3.4
langchain-community
pypdf==5.0.1
chromadb==0.5.15
# faiss-cpu==1.9.0